# hoteli-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hoteli API Dokumentacioni",
        "description": "Dokumentacioni i API-së për Sistemin e Rezervimeve të Hotelit",
        "contact": {
            "email": "support@hoteli.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/schedules/my": {
            "get": {
                "tags": [
                    "Schedules (Receptionist)"
                ],
                "summary": "Merr oraret e punës për përdoruesin e autentifikuar",
                "description": "Kthen oraret e punës për recepsionistin e autentifikuar ose të gjitha oraret nëse përdoruesi është admin. Kërkon autentifikim.",
                "operationId": "getMySchedules",
                "responses": {
                    "200": {
                        "description": "Orari i punës u mor me sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReceptionistSchedule"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Pa autentifikim",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Nuk jeni i autentikuar."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Gabim serveri",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Gabim gjatë marrjes së orareve."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/schedules/all": {
            "get": {
                "tags": [
                    "Schedules (Admin)"
                ],
                "summary": "Merr të gjitha oraret e punës",
                "description": "Kthen të gjitha oraret e punës për të gjithë recepsionistët. Vetëm adminët dhe recepsionistët mund ta aksesojnë.",
                "operationId": "getAllSchedules",
                "responses": {
                    "200": {
                        "description": "Lista e orareve u mor me sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReceptionistSchedule"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Pa autentifikim"
                    },
                    "403": {
                        "description": "Nuk keni leje",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Nuk keni leje për të parë të gjitha oraret."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Gabim serveri",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Gabim gjatë marrjes së orareve."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/schedules/{scheduleId}/status": {
            "put": {
                "tags": [
                    "Schedules (Receptionist)"
                ],
                "summary": "Përditëso statusin e një orari pune",
                "description": "Përditëson statusin e një orari specifik. Vetëm recepsionistët mund të ndryshojnë oraret e tyre.",
                "operationId": "updateScheduleStatus",
                "parameters": [
                    {
                        "name": "scheduleId",
                        "in": "path",
                        "description": "ID e orarit për të përditësuar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Statusi i ri i orarit",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Planned",
                                            "Completed",
                                            "Canceled"
                                        ],
                                        "example": "Completed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Statusi u përditësua me sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Statusi u përditësua me sukses."
                                        },
                                        "schedule": {
                                            "$ref": "#/components/schemas/ReceptionistSchedule"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Pa autentifikim"
                    },
                    "403": {
                        "description": "Nuk keni leje ose nuk mund të ndryshoni orarin e një recepsionisti tjetër",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Nuk keni leje për të ndryshuar statusin."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Orari nuk u gjet"
                    },
                    "422": {
                        "description": "Të dhëna të pavlefshme",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "status": [
                                                    "The selected status is invalid."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ReceptionistSchedule": {
                "title": "Receptionist Schedule",
                "description": "Detajet e orarit të një recepsionisti",
                "properties": {
                    "id": {
                        "description": "ID e orarit",
                        "type": "integer",
                        "readOnly": "true",
                        "example": 1
                    },
                    "user_id": {
                        "description": "ID e recepsionistit",
                        "type": "integer",
                        "example": 5
                    },
                    "shift_date": {
                        "description": "Data e turnit (YYYY-MM-DD)",
                        "type": "string",
                        "format": "date",
                        "example": "2024-06-15"
                    },
                    "start_time": {
                        "description": "Ora e fillimit të turnit (HH:MM:SS)",
                        "type": "string",
                        "format": "time",
                        "example": "08:00:00"
                    },
                    "end_time": {
                        "description": "Ora e mbarimit të turnit (HH:MM:SS)",
                        "type": "string",
                        "format": "time",
                        "example": "16:00:00"
                    },
                    "created_at": {
                        "description": "Data e krijimit",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "description": "Data e fundit e përditësimit",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
...
}